Parse CPU time values as u64 for Linux (closes #233)

+++ heim-common/src/sys/unix.rs
@@ -2,11 +2,11 @@
 
 use crate::{Error, Result};
 
-fn clock_ticks() -> Result<f64> {
+fn clock_ticks() -> Result<u64> {
     let result = unsafe { libc::sysconf(libc::_SC_CLK_TCK) };
 
     if result > 0 {
-        Ok(result as f64)
+        Ok(result as u64)
     } else {
         Err(Error::last_os_error().with_sysconf(libc::_SC_CLK_TCK))
     }
@@ -14,6 +14,6 @@ fn clock_ticks() -> Result<f64> {
 
 lazy_static::lazy_static! {
     /// Time units in USER_HZ or Jiffies
-    pub static ref CLOCK_TICKS: f64 = clock_ticks()
+    pub static ref CLOCK_TICKS: u64 = clock_ticks()
         .expect("Unable to determine CPU number of ticks per second");
 }
